System/am 002b event bus kafka topic (#37)
* fix(post-microservice): refactor database context configuration in services

Added a using directive for `Microsoft.EntityFrameworkCore`. Updated the `AddApplicationServices`
method to replace the previous PostgreSQL context addition with a new approach using `AddDbContext`
to configure the context with a connection string from the application's configuration. Retained the
old method as a commented reference.

* fix(post-service): update database connection handling in application service extensions

* gitStage(program.cs): AddAuthorization()

* feat(KafkaEvent): turnoff SecurityProtocol in KafkaConsumer config

This commit comments out the line that sets the `SecurityProtocol` to `Plaintext` in the `KafkaConsumer` configuration. The original line is retained as a comment for reference, effectively disabling the non-SSL configuration.

* feat(kafka topic): implement kafka topic register

---------

Co-authored-by: DuyKhiem <khiemnvdse180168@fpt.edu.vn>